Moved core link button functions to link.dart
