+++ lib/flutter_quill.dart
@@ -25,8 +25,7 @@ export 'src/utils/embeds.dart';
 export 'src/widgets/editor/editor.dart';
 export 'src/widgets/others/cursor.dart';
 export 'src/widgets/others/default_styles.dart';
-export 'src/widgets/others/link.dart'
-    show LinkActionPickerDelegate, LinkMenuAction;
+export 'src/widgets/others/link.dart';
 export 'src/widgets/quill/embeds.dart';
 export 'src/widgets/quill/quill_controller.dart';
 export 'src/widgets/raw_editor/raw_editor.dart';
